upgrade to v511

+++ adodb-memcache.lib.inc.php
@@ -11,7 +11,7 @@ if (empty($ADODB_INCLUDED_CSV)) include(ADODB_DIR.'/adodb-csvlib.inc.php');
 
 /* 
 
-  v4.991 16 Oct 2008  (c) 2000-2008 John Lim (jlim#natsoft.com). All rights reserved.
+  V5.06 16 Oct 2008  (c) 2000-2010 John Lim (jlim#natsoft.com). All rights reserved.
   Released under both BSD license and Lesser GPL library license. 
   Whenever there is any discrepancy between the two licenses, 
   the BSD license will take precedence. See License.txt. 
@@ -29,7 +29,7 @@ $db->memCacheCompress = false; /// Use 'true' to store the item compressed (uses
 
 $db->Connect(...);
 $db->CacheExecute($sql);
-
+  
   Note the memcache class is shared by all connections, is created during the first call to Connect/PConnect.
   
   Class instance is stored in $ADODB_CACHE
@@ -65,7 +65,7 @@ $db->CacheExecute($sql);
 
 			$memcache = new MemCache;
 			
-			if (!is_array($this->hosts)) $this->hosts = array($hosts);
+			if (!is_array($this->hosts)) $this->hosts = array($this->hosts);
 		
 			$failcnt = 0;
 			foreach($this->hosts as $host) {
@@ -77,13 +77,13 @@ $db->CacheExecute($sql);
 				$err = 'Can\'t connect to any memcache server';
 				return false;
 			}
-			
+			$this->_connected = true;
 			$this->_memcache = $memcache;
-			return 0;
+			return true;
 		}
 		
 		// returns true or false. true if successful save
-		function writecache($filename, $contents,$debug, $secs2cache)
+		function writecache($filename, $contents, $debug, $secs2cache)
 		{
 			if (!$this->_connected) {
 				$err = '';
@@ -91,7 +91,7 @@ $db->CacheExecute($sql);
 			}
 			if (!$this->_memcache) return false;
 			
-			if (!$this->_memcache->set($filename, $contents, $this->compress, 0)) {
+			if (!$this->_memcache->set($filename, $contents, $this->compress, $secs2cache)) {
 				if ($debug) ADOConnection::outp(" Failed to save data at the memcached server!<br>\n");
 				return false;
 			}
@@ -100,7 +100,7 @@ $db->CacheExecute($sql);
 		}
 		
 		// returns a recordset
-		function &readcache($filename, &$err, $secs2cache, $rsClass)
+		function readcache($filename, &$err, $secs2cache, $rsClass)
 		{
 			$false = false;
 			if (!$this->_connected) $this->connect($err);
@@ -111,16 +111,18 @@ $db->CacheExecute($sql);
 				$err = 'Item with such key doesn\'t exists on the memcached server.';
 				return $false;
 			}
-	
+			
 			// hack, should actually use _csv2rs
 			$rs = explode("\n", $rs);
             unset($rs[0]);
             $rs = join("\n", $rs);
-            $rs = unserialize($rs);
+ 			$rs = unserialize($rs);
 			if (! is_object($rs)) {
 				$err = 'Unable to unserialize $rs';		
 				return $false;
 			}
+			if ($rs->timeCreated == 0) return $rs; // apparently have been reports that timeCreated was set to 0 somewhere
+			
 			$tdiff = intval($rs->timeCreated+$secs2cache - time());
 			if ($tdiff <= 2) {
 				switch($tdiff) {
@@ -143,14 +145,14 @@ $db->CacheExecute($sql);
 			}
 			return $rs;
 		}
-	
+		
 		function flushall($debug=false)
 		{
 			if (!$this->_connected) {
 				$err = '';
 				if (!$this->connect($err) && $debug) ADOConnection::outp($err);
 			}
-			if ($this->_memcache) return false;
+			if (!$this->_memcache) return false;
 			
 			$del = $this->_memcache->flush();
 			
@@ -163,6 +165,12 @@ $db->CacheExecute($sql);
 		
 		function flushcache($filename, $debug=false)
 		{
+			if (!$this->_connected) {
+  				$err = '';
+  				if (!$this->connect($err) && $debug) ADOConnection::outp($err); 
+			} 
+			if (!$this->_memcache) return false;
+  
 			$del = $this->_memcache->delete($filename);
 			
 			if ($debug)
